refactor: rename vendor backend txda -> tsingmicro, unify vendor module resolution
- Rename sglang_fl/dispatch/backends/vendor/txda -> tsingmicro
- Update all internal import paths in patches, impl, and register modules
- Remove vendor name special-casing in sglang_fl/__init__.py and platform.py
- Update example scripts: fix import paths, add timeout env vars, move mem_fraction_static to Engine constructor
