+++ sglang_fl/dispatch/config/tsingmicro.yaml
@@ -6,6 +6,6 @@ prefer: flagos
 # Each op tries candidates left-to-right; falls to the next on failure.
 op_backends:
   silu_and_mul:
-    - vendor       # vendor.txda – pure PyTorch, safe on TXDA hardware
+    - vendor       # vendor.tsingmicro – pure PyTorch, safe on TXDA hardware
     - flagos       # default.flagos – Triton kernel, may not be TXDA-verified
     - reference
